refactor v1 and v2 in different packages
